pyproject.toml
[project]
name = "grav-font"
version = "0.1.0"
description = "Gravitational Font Tracing - Stylized glyph rendering using simulated 2D gravitational systems"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-image>=0.20.0",
    "matplotlib>=3.7.0",
    "freetype-py>=2.4.0",
    "skia-python>=87.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
grav-font = "run_demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long - let formatter handle it

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
